Reworked subscriptions handling (again)
Previous approach, even after fix, lacked from design problem:

let's say user A subscribed to channel A and started listening for
messages.
Subscription class for user A notices that nobody is listening now
and starts listening itself.

Then user B subscribes to channel B and starts listener.
Subscription class notices that user A is already listening
and just waits on its queue.

After that, let's say user A received a message.
His subscription class takes that message and stops listening.
But user B is still waiting on his queue,
and there is nobody to write to it.

So I had to add separate task
which will always listen for new messages on subscriptions
and put them to active queues.